Refactor publisher/subscriber as part of logcat.
