+++ mobly/controllers/android_device_lib/services/logcat.py
@@ -11,16 +11,27 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
+import collections
 import copy
+import datetime
+import fnmatch
 import io
 import logging
 import os
+import re
+import threading
 
+from dateutil.parser import parse as parse_date
 from mobly import logger as mobly_logger
 from mobly import utils
 from mobly.controllers.android_device_lib import adb
 from mobly.controllers.android_device_lib import errors
 from mobly.controllers.android_device_lib.services import base_service
+
+
+LogcatData = collections.namedtuple(
+    'LogcatData', ['time', 'pid', 'tid', 'level', 'tag', 'message',
+                   'host_time', 'line'])
 
 
 class Error(errors.ServiceError):
@@ -43,6 +54,10 @@ class Config(object):
 
 class Logcat(base_service.BaseService):
     """Android logcat service for Mobly's AndroidDevice controller."""
+    line_regex = re.compile(
+        r'(?P<time>\d\d-\d\d \d\d:\d\d:\d\d.\d\d\d)\s+'
+        r'(?P<pid>\d+)\s+(?P<tid>\d+)\s+(?P<level>[VDIWEFS])\s+'
+        r'(?P<tag>.+?)\s*:\s+(?P<message>.*)')
 
     def __init__(self, android_device, configs=None):
         super(Logcat, self).__init__(android_device, configs)
@@ -50,6 +65,9 @@ class Logcat(base_service.BaseService):
         self._adb_logcat_process = None
         self.adb_logcat_file_path = None
         self._configs = configs if configs else Config()
+        self._publisher_thread = None
+        self._publisher_process = None
+        self._subscribers = []
 
     def _enable_logpersist(self):
         """Attempts to enable logpersist daemon to persist logs."""
@@ -177,6 +195,8 @@ class Logcat(base_service.BaseService):
         process = utils.start_standing_subprocess(cmd, shell=True)
         self._adb_logcat_process = process
         self.adb_logcat_file_path = logcat_file_path
+        if self._subscribers:
+            self._start_publisher()
 
     def stop(self):
         """Stops the adb logcat service."""
@@ -187,6 +207,7 @@ class Logcat(base_service.BaseService):
         except:
             self._ad.log.exception('Failed to stop adb logcat.')
         self._adb_logcat_process = None
+        self._stop_publisher()
 
     def pause(self):
         """Pauses logcat for usb disconnect."""
@@ -208,3 +229,199 @@ class Logcat(base_service.BaseService):
         resume_configs = copy.copy(self._configs)
         resume_configs.clear_log = False
         self.start(resume_configs)
+
+    def subscribe(self, subscriber):
+        """Subscribe a subscriber to this publisher.
+
+        Args:
+            subscriber: LogcatSubscriber, a logcat subscriber to subscribe.
+
+        Raises:
+            Error: When subscriber is not a LogcatSubscriber.
+        """
+        if not self._publisher_is_active:
+            self._start_publisher()
+
+        if not isinstance(subscriber, LogcatSubscriber):
+            raise Error(self._ad, 'Attempt to subscribe a non-subscriber.')
+        self._subscribers.append(subscriber)
+
+    def unsubscribe(self, subscriber):
+        """Unsubscribe a subscriber to this publisher.
+
+        Args:
+            subscriber: LogcatSubscriber, a logcat subscriber to unsubscribe.
+
+        Raises:
+            Error: When the argument is not previously registered subscriber.
+        """
+        if subscriber not in self._subscribers:
+            raise Error(self._ad, 'Attempt to unsubscribe a non-subscriber.')
+        self._subscribers.remove(subscriber)
+
+    def event(self, pattern='.*', tag='*', level='V'):
+        """Context manager object for a logcat event.
+
+        Args:
+            pattern: str, Regular expression pattern to trigger on.
+            tag: str, Tag portion of filterspec string.
+            level: str, Level portion of filterspec string.
+
+        Returns:
+            A context manager describing the event.
+        """
+        subscriber = LogcatEventSubscriber(
+            pattern=pattern, tag=tag, level=level)
+        subscriber.subscribe(self)
+        return subscriber
+
+    @property
+    def _publisher_is_active(self):
+        """Returns a boolean value indicating if publisher is running."""
+        return (self._publisher_process
+                and self._publisher_process.returncode is None
+                and self._publisher_thread
+                and self._publisher_thread.is_alive())
+
+    def _start_publisher(self):
+        """Start the publisher process and task."""
+        if self.adb_logcat_file_path is None:
+            raise Error('ADB logcat file path is not defined.')
+        if self._publisher_is_active:
+            raise Error('Publisher process is already running.')
+
+        cmd = ['tail', '-F', self.adb_logcat_file_path]
+        self._publisher_process = utils.start_standing_subprocess(cmd)
+        self._publisher_thread = threading.Thread(
+            target=self._publisher_task)
+        self._publisher_thread.daemon = True
+        self._publisher_thread.start()
+
+    def _stop_publisher(self):
+        """Stop the publisher process and task."""
+        if self._publisher_is_active:
+            self._publisher_process.terminate()
+            self._publisher_process.wait()
+            self._publisher_thread.join()
+
+    def _publisher_task(self):
+        """Main publisher thread task."""
+        for line in iter(self._publisher_process.stdout.readline, ''):
+            match = self.line_regex.match(line)
+            if match is None:
+                continue
+            time = match and parse_date(match.group('time'))
+            pid = match and int(match.group('pid'))
+            tid = match and int(match.group('tid'))
+            level = match and match.group('level')
+            tag = match and match.group('tag')
+            message = match and match.group('message').strip()
+
+            pub_data = LogcatData(
+                time=time, pid=pid, tid=tid, level=level, tag=tag,
+                message=message, host_time=datetime.datetime.now(), line=line)
+
+            for subscriber in self._subscribers:
+                subscriber.handle(pub_data)
+
+
+class LogcatSubscriber(object):
+    """Base class for logcat subscriber."""
+
+    def __init__(self):
+        super(LogcatSubscriber, self).__init__()
+        self._publisher = None
+
+    def subscribe(self, publisher):
+        """Subscribe this object to a publisher.
+
+        Args:
+            publisher: LogcatPublisher, a logcat publisher to subscribe to.
+
+        Raises:
+            Error: If publisher is not a LogcatPublisher.
+        """
+        if not isinstance(publisher, Logcat):
+            raise Error('Only Logcat instance can be subscribed to.')
+        publisher.subscribe(self)
+        self._publisher = publisher
+
+    def unsubscribe(self):
+        """Unsubscribe this object to its publisher."""
+        self._publisher.unsubscribe(self)
+        self._publisher = None
+
+    def handle(self, data):
+        """Abstract subscribe handler method.
+
+        This abstract method defines the subscribe handler and must be
+        overridden by derived class.
+
+        Args:
+            data: LogcatData, Data to handle.
+        """
+        raise NotImplementedError('"handle" is a required subscriber method.')
+
+
+class LogcatEventSubscriber(LogcatSubscriber):
+    """Logcat event subscriber class.
+
+    This class waits for a particular logcat event to occur.
+
+    Args:
+        pattern: str, Regular expression pattern to trigger on.
+        tag: str, Tag portion of filterspec string.
+        level: str, Level portion of filterspec string.
+    """
+    _LOG_LEVELS = 'VDIWEF'
+
+    def __init__(self, pattern='.*', tag='*', level='V'):
+        super(LogcatEventSubscriber, self).__init__()
+        self._event = threading.Event()
+        self._pattern = (re.compile(pattern) if isinstance(pattern, str)
+                         else pattern)
+        self._tag = tag
+        self._levels = (self._LOG_LEVELS if level == '*'
+                        else self._LOG_LEVELS[self._LOG_LEVELS.find(level):])
+        self.trigger = None
+        self.match = None
+
+    def __enter__(self):
+        return self
+
+    def __exit__(self, exc_type, exc_value, traceback):
+        self.unsubscribe()
+
+    def clear(self):
+        """Clears any set event."""
+        self.trigger = None
+        self.match = None
+        self._event.clear()
+
+    def wait(self, timeout=None):
+        """Wait until the trigger expression is seen.
+
+        Args:
+            timeout: float, Timeout in seconds.
+
+        Returns:
+            True except if a timeout is given and the operation times out.
+        """
+        return self._event.wait(timeout)
+
+    def handle(self, data):
+        """Handle a logcat subscription message.
+
+        Args:
+            data: LogcatData, Data to handle.
+        """
+        if self.trigger:
+            return
+        if data.tag is None or not fnmatch.fnmatchcase(data.tag, self._tag):
+            return
+        if data.level is None or data.level not in self._levels:
+            return
+        self.match = self._pattern.match(data.message)
+        if self.match:
+            self.trigger = data
+            self._event.set()
